#issue 43: more tests.

+++ test/CodeGen/EVM/runtime_tests/bitwise/swap_bits.c
@@ -1,10 +1,10 @@
 long  swap(long n, long p, long q)
 {
 	// see if the bits are same. we use XOR operator to do so.
-	if (((n & (1 << p)) >> p) ^ ((n & (1 << q)) >> q))
+	if (((n & ((long)1 << p)) >> p) ^ ((n & ((long)1 << q)) >> q))
 	{
-		n ^= (1 << p);
-		n ^= (1 << q);
+		n ^= ((long)1 << p);
+		n ^= ((long)1 << q);
 	}
  
 	return n;
